+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.6.1
+    - fix `output` decorator, it will modify schema_json's `required` field and fill all fields
+
 ## v1.6.0
 
 - add `Resolver.context` param